# perf: `take_run` improvements